bug fixed

+++ Aban360.ReportPool.Domain/Features/ConsumersInfo/Dto/PreviousConsumptionsDto.cs
@@ -2,7 +2,7 @@
 {
     public record PreviousConsumptionsDto
     {
-        public int ConsumptionAmount { get; set; }
+        public long  ConsumptionAmount { get; set; }
         public string ConsumptionDateJalali { get; set; }
     }
 }
